[CHORE] adding react router
