+++ ui/package.json
@@ -31,6 +31,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-query": "^3.39.3",
+    "react-router-dom": "^6.28.0",
     "sql-formatter": "^15.4.5",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7"
