+++ ui/src/main.tsx
@@ -5,6 +5,7 @@ import App from './App.tsx'
 import { ThemeProvider } from './theme-provider.tsx'
 import { QueryClient, QueryClientProvider } from 'react-query'
 import { Toaster } from './components/ui/toaster.tsx'
+import { BrowserRouter } from 'react-router-dom'
 
 const queryClient = new QueryClient();
 
@@ -12,8 +13,10 @@ createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <QueryClientProvider client={queryClient}>
       <ThemeProvider defaultTheme="dark" storageKey="vite-ui-theme">
-        <App />
-        <Toaster />
+        <BrowserRouter>
+          <App />
+          <Toaster />
+        </BrowserRouter>
       </ThemeProvider>
     </QueryClientProvider>
   </StrictMode>,
